project closed, not perfect, but that's it.

+++ 1.Firmware/TouchGFX/gui/include/gui/details_screen/DetailsView.hpp
@@ -11,7 +11,7 @@ public:
     virtual void NH3Update(float adc);
     virtual void SHT31Update(float value1,float value2);
     virtual void BH1750Update(float value);
-    virtual void CO2Upadate(uint32_t co2);
+    virtual void CO2Upadate(uint16_t co2);
     virtual ~DetailsView() {}
     virtual void setupScreen();
     virtual void tearDownScreen();
